Move all aiohttp-related things to `microchat/app`
