+++ microchat/app/api_adapters/misc.py
